backend

+++ crates/defi/src/types.rs
@@ -33,7 +33,7 @@ impl From<Risk> for String{
 #[derive(Debug,Serialize, Deserialize, Clone)]
 pub struct Asset{
     name: String,
-    balance: u32
+    balance: f64
 }
 
 impl From<Asset> for String{
